packages/knowledge_base/requirements.txt
# =============================================================================
# GOOGLE CLOUD - Core
# =============================================================================
google-cloud-storage>=2.18.0
google-cloud-discoveryengine>=0.11.0

# =============================================================================
# FILE PROCESSING
# =============================================================================
PyPDF2>=3.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
markdown>=3.5.0

# =============================================================================
# UTILITIES
# =============================================================================
tenacity>=8.2.0  # Retry logic
